Added more comments and created HPCToolKitReader class
