+++ pipit/readers/hpctoolkit_reader.py
@@ -1,0 +1,298 @@
+# Copyright 2017-2021 Lawrence Livermore National Security, LLC and otherfrom xml.etree.ElementTree import ElementTree
+from asyncio.format_helpers import _get_function_source
+from xml.etree.ElementTree import Element, ElementTree
+import pandas as pd
+
+from Graph import Graph, Node
+
+
+
+
+
+		
+
+		
+
+
+
+
+	
+		
+
+
+
+class ExperimentReader:
+	
+	
+	def __init__(self, file_location):
+		self.tree = ElementTree(file = file_location)
+
+	def get_function_name(self, metric_id):
+
+		#first get id in procedure table
+		search = './/S[@it=\"' + str(metric_id) + '\"]..'
+		e = self.tree.find(search)
+		procedure_table_id = e.get('n')
+		return self.get_function_name_helper(procedure_table_id)
+
+		
+
+	def get_function_name_helper(self, procedure_table_id):
+		#return function name
+		procedure_table_search = './/Procedure[@i=\'' + procedure_table_id + '\']'
+		procedure = self.tree.find(procedure_table_search)
+		return procedure.get('n')
+
+	def get_min_max_time(self):
+		search = './/TraceDB[@i=\"0\"]'
+		e = self.tree.find(search)
+		time = (int(e.get('db-min-time')), int(e.get('db-max-time')))
+		return time
+
+	def create_graph(self):
+		graph = Graph()
+		call_path_data = list(list(self.tree.getroot())[-1])[-1]
+		root_elems = list(call_path_data)
+		for root_elem in root_elems:
+			# procedure_table_id = root_elem.attrib['n']
+			# function_name = self.get_function_name_helper(procedure_table_id)
+			# node = Node(procedure_table_id, function_name, None)
+			node = self.graph_helper(None, root_elem, graph)
+			if node != None:
+				graph.add_root(node)	
+
+		return graph
+
+	def graph_helper(self, parent_node: Node, curr_element: Element, graph: Graph):
+		if(curr_element.tag == 'PF'):
+			procedure_table_id = curr_element.attrib['n']
+			function_name = self.get_function_name_helper(procedure_table_id)
+			new_node = Node(procedure_table_id, function_name, parent_node)
+			if(parent_node != None):
+				parent_node.add_child(new_node)
+
+			for child_elem in list(curr_element):
+				self.graph_helper(new_node, child_elem, graph)
+			return new_node
+		
+		else:
+			calling_context_id = curr_element.attrib.get('it')
+			if(calling_context_id != None):
+				parent_node.add_calling_contex_id(calling_context_id)
+				graph.add_to_map(calling_context_id, parent_node)
+
+			for child_elem in list(curr_element):
+				self.graph_helper(parent_node, child_elem, graph)
+			return None
+			
+class ProfileReader:
+
+
+	def __init__(self, file_location):
+		self.file = open(file_location, "rb")
+		file = self.file
+		file.seek(32)
+		
+		# need to test to see if correct
+		byte_order = 'big'
+		signed = False
+
+
+
+		# Profile Info section offset (pi_ptr)
+		
+		self.pi_ptr = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+
+	def read_info(self, prof_info_idx):
+		byte_order = 'big'
+		signed = False
+		file = self.file
+		
+
+		# Profile Info
+		file.seek(self.pi_ptr + (prof_info_idx * 52))
+		idt_ptr = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+		
+		#skipping because not in use 
+		# file.read(24) 
+
+		# num_vals = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+		# num_nzctxs = int.from_bytes(file.read(4), byteorder=byte_order, signed=signed)
+		# prof_off = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+
+
+		# Hierarchical Identifier Tuple
+		file.seek(idt_ptr)
+		num_tuples = int.from_bytes(file.read(2), byteorder=byte_order, signed=signed)
+		# print('num_tuples', num_tuples)
+		tuples_list = []
+		for i in range(0, num_tuples, 1):
+			# not working --  I don't know why, but the second 2 tuples are just incorrect
+			kind = int.from_bytes(file.read(2), byteorder=byte_order, signed=signed)
+			p_val = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+			l_val = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+			tuples_list.append((kind, p_val, l_val))
+		return tuples_list
+
+
+class HPCToolkitReader:
+	
+
+	def __init__(self, dir_name) -> None:
+		self.dir_name = dir_name # directory of hpctoolkit trace files being read
+
+	def read(self):
+		# this function reads through the trace.db file with two nested loops - 
+		# the outer loop iterates on the rank of the process while the inner loop 
+		# iterates through each trace line for the fiven rank and adds it to the dictionary
+		
+		dir_location = self.dir_name # directory of hpctoolkit trace files being read
+
+		# open file
+		file = open(dir_location + "trace.db", "rb")
+
+		experiment_reader = ExperimentReader(dir_location + "experiment.xml")
+
+		profile_reader = ProfileReader(dir_location + "profile.db")
+
+		# create graph
+		graph = experiment_reader.create_graph()
+		
+		# read Magic identifier ("HPCPROF-tracedb_")
+		encoding = 'ASCII' #idk just guessing rn
+		identifier = str(file.read(16), encoding)
+		
+		# read version
+		version_major = file.read(1)
+		version_minor = file.read(1)
+
+		# need to test to see if correct
+		byte_order = 'big'
+		signed = False
+
+		# Number of trace lines (num_traces)
+		num_traces = int.from_bytes(file.read(4), byteorder=byte_order, signed=signed)
+		# print("num_traces", num_traces)
+
+		# Number of sections (num_sec)
+		num_sections = int.from_bytes(file.read(2), byteorder=byte_order, signed=signed)
+		# print("num_sections", num_sections)
+
+		# Trace Header section size (hdr_size)
+		hdr_size = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+		
+		# Trace Header section offset (hdr_ptr)
+		hdr_ptr = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+		# print("hdr_size: ", hdr_size)
+
+		data = {'Function Name':[], 'Event Type':[], 'Time':[], 'Process':[], 'Graph_Node': [], 'Level': []}
+		min_max_time = experiment_reader.get_min_max_time()
+
+
+		# cycle through trace headers (each iteration in this outer loop is a seperate process/thread/rank)
+		for i in range(0, hdr_size, 22):
+			proc_num = int(i/22)
+			file.seek(hdr_ptr + i)
+			
+			# prof_info_idx (in profile.db)
+			prof_info_idx = int.from_bytes(file.read(4), byteorder=byte_order, signed=signed)
+			
+
+			# Trace type
+			trace_type = int.from_bytes(file.read(2), byteorder=byte_order, signed=signed)
+
+			# Offset of Trace Line start (line_ptr) 
+			line_ptr = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+			
+			# Offset of Trace Line one-after-end (line_end) 
+			line_end = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed)
+
+			last_id = -1
+
+			last_node = None
+
+
+# iterate through every trace event in the process 
+			for j in range (line_ptr, line_end, 12):
+				file.seek(j)
+			
+				# Timestamp (nanoseconds since epoch)
+				timestamp = int.from_bytes(file.read(8), byteorder=byte_order, signed=signed) - min_max_time[0]
+				
+				# Sample calling context id (in experiment.xml)
+				# can use this to get name of function from experiement.xml Procedure tab
+				calling_context_id = int.from_bytes(file.read(4), byteorder=byte_order, signed=signed)
+
+
+# checking to see if the last event wasn't the same 
+# if it was then we skip it, as to not have multiple sets of open/close events for a function that it's still in 					
+				if(last_id != calling_context_id):
+
+					# updating the trace_db
+
+					node = graph.get_node(calling_context_id) # the node in the CCT
+					
+					# closing functions exited
+					close_node = last_node
+					intersect_level = -1
+					intersect_node = node.get_intersection(last_node) 
+					# this is the highest node that last_node and node have in common 
+					# we want to close every enter time event higher than than interest node, because those functions have exited
+					
+					
+					if(intersect_node != None):
+						intersect_level = intersect_node.get_level()	
+					while close_node != None and close_node.get_level() > intersect_level:
+						data['Function Name'].append(close_node.name)
+						data['Event Type'].append('Exit')
+						data['Time'].append(timestamp)
+						data['Process'].append(proc_num)
+						data['Graph_Node'].append(close_node)
+						data['Level'].append(close_node.get_level())
+						close_node = close_node.parent
+					
+
+					
+					# creating new rows for the new functions entered 
+					enter_list = node.get_node_list(intersect_level) 
+					# the list of nodes higher than interesect_level (the level of interesect_node)
+					# all of the nodes in this list have entered into the function since the last poll
+					# so we want to create entries in the data for the Enter event
+					for enter_node in enter_list[::-1]:
+						data['Function Name'].append(enter_node.name)
+						data['Event Type'].append('Enter')
+						data['Time'].append(timestamp)
+						data['Process'].append(proc_num)
+						data['Graph_Node'].append(enter_node)
+						data['Level'].append(enter_node.get_level())
+					last_node = node
+						
+				last_id = calling_context_id # updating last_id
+
+
+			# adding last data for trace df
+			close_node = last_node
+
+			# after reading through all the trace lines, some Enter events will not have matching Exit events,
+			# as the functions were still running in the last poll.
+			# Here we are adding Exit events to all of the remaining unmatched Enter events 
+			while close_node != None:
+				data['Function Name'].append(close_node.name)
+				data['Event Type'].append('Exit')
+				data['Time'].append(min_max_time[1] - min_max_time[0])
+				data['Process'].append(proc_num)
+				data['Graph_Node'].append(close_node)
+				data['Level'].append(close_node.get_level())
+				close_node = close_node.parent
+
+		# primitive_df doesn't have full trace, but has just the top level function at all times
+		trace_df = pd.DataFrame(data) 
+		self.trace_df = trace_df
+		return pipit.tracedata.TraceData(trace_df)
+
+				
+
+
+		
+	
+
